requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.31.0
PyPDF2==3.0.1
pdfplumber==0.10.3
pandas==2.2.1
matplotlib==3.8.3
seaborn==0.13.2
Pillow==10.2.0
numpy==1.26.4
torch==2.2.1
transformers==4.38.2
sentence-transformers==2.5.1
scikit-learn==1.4.1.post1
opencv-python==4.9.0.80
PyMuPDF==1.23.26
nltk==3.8.1
networkx==3.2.1
faiss-cpu==1.7.4
langchain
langchain-community
langchain-core
chromadb
tiktoken